+++ muddler_project/src/triggers/Levi_Ataxia/Bard_Slash_2.lua
@@ -1,0 +1,11 @@
+local weapons = {"Soulpiercer", "Eagle's Scream","rapier"}
+
+for i in pairs(weapons) do
+ if string.find(multimatches[1][3], weapons[i]) then
+ 
+	slc_last_limb = multimatches[1][4]
+	
+	weapcheck = true
+	tempTimer(1, [[weapcheck = false]] )
+ end
+end