+++ muddler_project/src/triggers/Levi_Ataxia/Fullsense-Demon.lua
@@ -1,0 +1,4 @@
+capturing_fullsense_demon = true
+mmp.pdb_lastupdate = {}
+
+fullSensePeople = {}