+++ muddler_project/src/triggers/Levi_Ataxia/Fullsense-Hyena.lua
@@ -1,0 +1,4 @@
+capturing_fullsense_hyena = true
+mmp.pdb_lastupdate = {}
+
+fullSensePeople = {}