+++ muddler_project/src/triggers/Levi_Ataxia/GMCP_Map_Catchers.lua
@@ -1,0 +1,3 @@
+if not zgui.map then
+  zgui.buildMap()
+end