+++ muddler_project/src/triggers/Levi_Ataxia/Get_Humour_Levels.lua
@@ -1,0 +1,11 @@
+deleteLine()
+cecho("\n<a_blue> --- [ <NavajoWhite>EVALUATE: <a_green>"..target:upper().."<a_blue> ] ---")
+tEval = {
+	sanguine = 0, choleric = 0, phlegmatic = 0, melancholic = 0,
+	hp = 0, mp = 0,
+}
+
+tAffs.sanguine = nil
+tAffs.melancholic = nil
+tAffs.phlegmatic = nil
+tAffs.choleric = nil