+++ muddler_project/src/triggers/Levi_Ataxia/Hiraku_2.lua
@@ -1,0 +1,1 @@
+slc_last_limb = "head"