+++ muddler_project/src/triggers/Levi_Ataxia/Readaura_stuff.lua
@@ -1,0 +1,2 @@
+capturing_readaura = true
+deleteLine()