+++ muddler_project/src/triggers/Levi_Ataxia/Room_Info_Shortener.lua
@@ -1,0 +1,8 @@
+local x = ataxia.settings.roomshorten
+
+if gmcp.Room.Info.name:lower():find(matches[2]:lower()) then
+	if (x == "notmanual" and ataxiaBasher.manual ~= true) or (x == "normal") then
+		roomDeleting = true
+		deleteLine()
+	end
+end